third_party/blink/web_tests/platform/generic/fast/js/mozilla/strict/15.4.5.1-expected.txt

PASS 'use strict'; var a = arr(); a.length = 2; a threw exception of type TypeError.
PASS true === true
PASS 'use strict'; var a = addx(arr()); a.length = 2; a threw exception of type TypeError.
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
